Improved and fixed detections
- Improved CommandCheck
Now it will check for its arguments if a command should be blocked. If a "*" is put at the end, it will check if the command starts with the string provided.
- Probably fixed detections in commands and chats
Probably related to #15
The event is called and the sanction is considered, but the respective actions are not performed.
Needs to be tested in production before releasing update